Refactored code from yesterday
I moved the Header struct back into its own mod.  I also cleaned up the
logic for grabbing the headers to use a draining iterator on the bytes
in the file.  I think this is a more elegant solution for getting the
bytes than keeping track of the offset myself.