Adds copyright header to new files
The biggest change is that I added the copyright header to both the
Archive and the Header module files.  Since the copyright that I've
given this requires the header, it seemed like a good thing to do.

I also went through the code and cleaned it up so I wouldn't get as many
warnings when compiling.